requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.46
PyJWT==2.8.0
psycopg2-binary==2.9.9
